+++ recording/src/main/java/mc/replay/recording/file/RecordingFormat.java
@@ -9,6 +9,6 @@ public final class RecordingFormat {
 
     public static final String FILE_EXTENSION = ".mcrr"; // MC-Replay Recording
 
-    public static final int RECORDING_VERSION = 1;
-    public static final int MINECRAFT_PROTOCOL_VERSION = ProtocolVersion.getServerVersion().getNumber();
+    public static final byte RECORDING_VERSION = 1;
+    public static final byte MINECRAFT_PROTOCOL_VERSION = (byte) ProtocolVersion.getServerVersion().getNumber();
 }
